+++ test/i18nextProvider.spec.js
@@ -17,13 +17,4 @@ describe('I18nextProvider', () => {
       '[react-i18next][I18nextProvider]does not support changing the i18n object.'
     );
   });
-  it('should have i18n proptype required', () => {
-    expect(I18nextProvider.propTypes.i18n).toBe(PropTypes.object.isRequired);
-  });
-  it('should have defaultNS proptype optional string', () => {
-    expect(I18nextProvider.propTypes.defaultNS).toBe(PropTypes.string);
-  });
-  it('should have children proptype required', () => {
-    expect(I18nextProvider.propTypes.children).toBe(PropTypes.element.isRequired);
-  });
 });